+++ src/CommunityPatchLauncher/ViewModels/MainWindowModel.cs
@@ -1,11 +1,15 @@
 ﻿using CommunityPatchLauncher.BindingData.Container;
 using CommunityPatchLauncher.Commands.ApplicationWindow;
+using CommunityPatchLauncher.Commands.Os;
+using CommunityPatchLauncher.Documentation.Factories;
 using CommunityPatchLauncher.Enums;
 using CommunityPatchLauncher.Settings.Factories;
 using CommunityPatchLauncher.UserControls;
 using CommunityPatchLauncherFramework.Settings.Factories;
 using CommunityPatchLauncherFramework.Settings.Manager;
 using System;
+using System.IO;
+using System.Reflection;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
@@ -17,6 +21,11 @@ namespace CommunityPatchLauncher.ViewModels
     /// </summary>
     internal class MainWindowModel : BaseViewModel
     {
+        /// <summary>
+        /// Open the news to the window
+        /// </summary>
+        public ICommand OpenNewsCommand { get; private set; }
+
         /// <summary>
         /// The command used to open the changelog
         /// </summary>
@@ -58,6 +67,16 @@ namespace CommunityPatchLauncher.ViewModels
         public ICommand ReportIssueCommand { get; private set; }
 
         /// <summary>
+        /// Open the s4 editor
+        /// </summary>
+        public ICommand OpenEditorCommand { get; private set; }
+
+        /// <summary>
+        /// Oopen the texture changer tool
+        /// </summary>
+        public ICommand OpenTextureChangerCommand { get; private set; }
+
+        /// <summary>
         /// The content dock to use
         /// </summary>
         private readonly DockPanel contentDock;
@@ -77,6 +96,8 @@ namespace CommunityPatchLauncher.ViewModels
             updateSearched = false;
             CloseWindowCommand = new CloseApplicationCommand();
 
+            SetWindowTitle();
+
             object dockArea = window.FindName("DP_ContentDock");
             if (dockArea is DockPanel panel)
             {
@@ -85,6 +106,7 @@ namespace CommunityPatchLauncher.ViewModels
 
                 contentDock = panel;
 
+                OpenNewsCommand = new OpenControlToPanel(contentDock, new BrowserUserControl("News.md", new RemoteDocumentManagerFactory(new TimeSpan(0, 30, 0))));
                 LaunchGameCommand = new OpenControlToPanel(contentDock, new PatchVersionSelectionUserControl(window));
                 OpenSettingCommand = new OpenControlToPanel(contentDock, new SettingsUserControl(currentWindow));
                 OpenChangelogCommand = new OpenControlToPanel(contentDock, new BrowserUserControl("Changelog.md"));
@@ -92,6 +114,22 @@ namespace CommunityPatchLauncher.ViewModels
                 OpenAboutCommand = new OpenControlToPanel(contentDock, new BrowserUserControl("About.md"));
                 ReportIssueCommand = new OpenLinkCommand(wpfSettings.GetValue<string>("ReportIssueLink"));
                 ComingSoonCommand = new OpenControlToPanel(contentDock, new ComingSoonControl());
+                object titleBarObject = currentWindow.FindName("TitleBar");
+                if (titleBarObject is TitleBarUseControl titleBar)
+                {
+                    titleBar.MouseDoubleClick += (sender, data) =>
+                    {
+                        MaximizeWindowCommand?.Execute(null);
+                    };
+                }
+
+                OpenEditorCommand = new StartEditorCommand(settingManager);
+
+                string gameFolder = settingManager.GetValue<string>("GameFolder");
+                string textureChange = gameFolder + "Texturenwechsler.bat";
+                OpenTextureChangerCommand = new StartProgramCommand(textureChange);
+
+                OpenNewsCommand.Execute(null);
             }
 
             ChangeGroupVisiblity = new ToggleSubGroupVisibilityCommand(currentWindow);
@@ -104,7 +142,25 @@ namespace CommunityPatchLauncher.ViewModels
                 updateSearched = true;
                 CheckForUpdateIfNeeded(window);
             };
+        }
 
+        /// <summary>
+        /// Add the version to the window title
+        /// </summary>
+        private void SetWindowTitle()
+        {
+            string version = settingManager.GetValue<string>("LauncherVersion");
+            Assembly assembly = Assembly.GetExecutingAssembly();
+            string versionString = "0.0.0";
+            using (Stream stream = assembly.GetManifestResourceStream("CommunityPatchLauncher.Version.txt"))
+            {
+                using (StreamReader reader = new StreamReader(stream))
+                {
+                    versionString = reader.ReadLine();
+                }
+            }
+            version = version == null ? versionString : "Unstable " + version;
+            WindowTitle += " - Version " + version;
         }
 
         /// <summary>